+++ src/main/java/application/domain/prices_updater/IPricesUpdater.java
@@ -1,8 +1,9 @@
 package application.domain.prices_updater;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
 public interface IPricesUpdater {
-    void updatePrices(List<Map.Entry<String,Float>> data, String invoiceDate);
+    void updatePrices(List<Map.Entry<String, BigDecimal>> data, String invoiceDate);
 }
